moved navbar to app

+++ Repositories/KeepRepository.cs
@@ -46,9 +46,9 @@ namespace keepr.Repositories
       }
     }
 
-    public bool Delete(int id)
+    public bool Delete(int id, string userId)
     {
-      int success = _db.Execute("DELETE FROM keeps WHERE id = @id", new { id });
+      int success = _db.Execute("DELETE FROM keeps WHERE id = @id AND userId = @userId", new { id, userId });
       return success > 0;
     }
